backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pip==23.3.1
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==68.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
wheel==0.41.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==23.4